Added Resources.sc and atND.sc
Resources increases SC resources through its *initClass, meaning that
this is done when the class library is compiled. It can be deactivated
by being moved to an 'ignore' folder.

atND and putND are additional methods to Array, that allow getting and
setting elements at variable depths (for multilevel arrays). The 'indexes'
argument is a sequence, meaning that it can be of variable length (which
is different from a fixed sequence of .at).